Changing dtypes to torch.long

+++ datasets/hm_visual_bert.py
@@ -44,9 +44,9 @@ class HMVisualBertDataset(Dataset):
         # torch.tensor automatically creates a copie of the data
         sample['img_features'] = torch.tensor(self.df['img_features'][idx], dtype=torch.float)
 
-        sample['input_ids'] = torch.tensor(self.df['text_encoding'][idx]['input_ids'].astype(np.int64), dtype=torch.int64)
-        sample['segment_ids'] = torch.tensor(self.df['text_encoding'][idx]['segment_ids'].astype(np.int8), dtype=torch.int8)
-        sample['input_mask'] = torch.tensor(self.df['text_encoding'][idx]['input_mask'].astype(np.int8), dtype=torch.int8)
+        sample['input_ids'] = torch.tensor(self.df['text_encoding'][idx]['input_ids'].astype(np.int64), dtype=torch.long)
+        sample['segment_ids'] = torch.tensor(self.df['text_encoding'][idx]['segment_ids'].astype(np.int64), dtype=torch.long)
+        sample['input_mask'] = torch.tensor(self.df['text_encoding'][idx]['input_mask'].astype(np.int64), dtype=torch.long)
 
         if self.name == 'test':
             sample['label'] = torch.tensor(np.array([]), dtype=torch.float)
@@ -57,4 +57,4 @@ class HMVisualBertDataset(Dataset):
 
     def getIdNumber(self, idx):
         """ Returns the id number of the image corresponding to index """
-        return torch.tensor(self.df['id'][idx], dtype=torch.int)
+        return torch.tensor(self.df['id'][idx], dtype=torch.long)
